+++ kamehouse-media/src/main/java/com/nicobrest/kamehouse/media/video/model/kamehousecommand/GetPlaylistContentKameHouseCommand.java
@@ -1,10 +1,8 @@
 package com.nicobrest.kamehouse.media.video.model.kamehousecommand;
 
-import com.nicobrest.kamehouse.commons.exception.KameHouseInvalidCommandException;
 import com.nicobrest.kamehouse.commons.model.kamehousecommand.KameHouseShellScript;
-import com.nicobrest.kamehouse.commons.utils.DockerUtils;
-import com.nicobrest.kamehouse.commons.utils.FileUtils;
 import com.nicobrest.kamehouse.commons.utils.StringUtils;
+import com.nicobrest.kamehouse.commons.validator.InputValidator;
 import java.util.List;
 
 /**
@@ -23,8 +21,8 @@ public class GetPlaylistContentKameHouseCommand extends KameHouseShellScript {
    */
   public GetPlaylistContentKameHouseCommand(String playlistFilename) {
     super();
-    validateFilename(playlistFilename);
-    this.playlistFilename = playlistFilename.replaceAll("\\\\", "/");
+    InputValidator.validateForbiddenCharsForShell(playlistFilename);
+    this.playlistFilename = playlistFilename.replace("\\", "/");
   }
 
   @Override
@@ -58,33 +56,5 @@ public class GetPlaylistContentKameHouseCommand extends KameHouseShellScript {
       return null;
     }
     return "-f " + playlistFilename;
-  }
-
-  /**
-   * Validate filename parameter.
-   */
-  private void validateFilename(String filename) {
-    if (StringUtils.isEmpty(filename)) {
-      throw new KameHouseInvalidCommandException("No file specified");
-    }
-    if (FileUtils.isRemoteFile(filename) || DockerUtils.shouldExecuteOnDockerHost(
-        executeOnDockerHost())) {
-      validateRemoteFile(filename);
-    } else {
-      if (!FileUtils.isValidLocalFile(filename)) {
-        throw new KameHouseInvalidCommandException(
-            "File doesn't exist on the server: " + filename);
-      }
-    }
-  }
-
-  /**
-   * Throw an exception if the filename contains unauthorized characters that could allow remote
-   * code execution.
-   */
-  private static void validateRemoteFile(String filename) {
-    if (StringUtils.isEmpty(filename)) {
-      throw new KameHouseInvalidCommandException("Empty file provided to vlc start");
-    }
   }
 }